ctrl s

+++ FModel/Views/Snooper/Cube.cs
@@ -63,7 +63,7 @@ public class Cube : Model
         };
 
         Sections = new Section[1];
-        Sections[0] = new Section(0, (uint) Indices.Length, 0, unrealMaterial);
+        Sections[0] = new Section(0, Indices.Length, 0, unrealMaterial);
 
         AddInstance(Transform.Identity);
     }
